feat!: remove a lot of animations and optimize for speed
