+++ src/components/arrow-title.tsx
@@ -1,12 +1,8 @@
-"use client";
-
 import { cn } from "@/lib/utils";
 import Image from "next/image";
-import RevealOnScroll from "./reveal-on-scroll";
 
 export default function ArrowTitle({
   text,
-  slideDirection,
   className,
   id,
   priority,
@@ -18,20 +14,7 @@ export default function ArrowTitle({
   priority?: boolean;
 }) {
   return (
-    <RevealOnScroll
-      className={cn("group flex gap-4 md:gap-12", className)}
-      variants={{
-        hidden: {
-          opacity: 0,
-          y: slideDirection === "bottom" ? -128 : 128,
-        },
-        visible: {
-          opacity: 1,
-          y: 0,
-        },
-      }}
-      id={id}
-    >
+    <div className={cn("group flex gap-4 md:gap-12", className)} id={id}>
       <Image
         src="/images/arrow.svg"
         alt="Chevron arrows pointing down"
@@ -51,6 +34,6 @@ export default function ArrowTitle({
         priority={priority}
         className="group-hover:-translate-y-4 duration-500 ease-out"
       />
-    </RevealOnScroll>
+    </div>
   );
 }
