+++ src/components/sections/experience.tsx
@@ -1,6 +1,5 @@
 import { cn } from "@/lib/utils";
 import Blob from "../blob";
-import RevealOnScroll from "../reveal-on-scroll";
 import { type Card, FocusCards } from "../ui/focus-cards";
 import { Timeline } from "../ui/timeline";
 import Wrapper from "../wrapper";
@@ -140,18 +139,11 @@ export default function Experience({ className }: { className?: string }) {
   ];
 
   return (
-    <RevealOnScroll
-      className={cn("relative", className)}
-      variants={{
-        hidden: { opacity: 0, y: 100 },
-        visible: { opacity: 1, y: 0 },
-      }}
-      once
-    >
+    <div className={cn("relative", className)}>
       <Wrapper>
         <Timeline data={data} />
       </Wrapper>
       <Blob />
-    </RevealOnScroll>
+    </div>
   );
 }
